requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
dnspython==2.2.1
email-validator==1.2.1
emailvalidator==0.3
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
passlib==1.7.4
pydantic==1.9.0
PyJWT==2.4.0
python-decouple==3.6
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
